requirements.txt
fastapi
uvicorn[standard]
gunicorn
jinja2
python-dotenv
redis
celery
httpx
pydantic